refactored RigidTransform converter
